(bugfix)fix a bug due to `glm::translate` position
In glm, we make each transform at the right of the matrix, so in the
code:
	trans = glm::scale(trans, glm::vec3(0.6, 0.6, 0.6));
	trans = glm::rotate(trans, static_cast<float>(time_value), glm::vec3(0.0, 0.0, 1.0));
	trans = glm::translate(trans, glm::vec3(std::cos(time_value), std::sin(time_value), 0.0f));
we predict the matrix of glm::scale is `A`, glm::rotate is `B` and so
on. The `trans` actually equals to `ABC`. When we left multiply `trans`
by the vertex postion, we do glm::translate at first and do glm::scale
and glm::rotate later, but the axis of the rotate is (0,0,1), it leads
to wrong result.